add template for pool alloction computations
